更新电机驱动程序以及编码器驱动程序

+++ QRobot_STM32/QRobot_FreeRTOS/User/motor/inc/motor.h
@@ -1,7 +1,7 @@
 #ifndef __MOTOR_H_
 #define __MOTOR_H_
 
-void Motor_Control(float Motor1, float Motor2, float Motor3, float Motor4);
+void Motor_Control(int Motor1, int Motor2, int Motor3, int Motor4);
 void Motor_Stop(void);
 
 #endif
